requirements.txt
cachetools==5.3.1
click==8.1.3
crc==4.2.0
spidev==3.6; sys_platform=="linux"
RPi.GPIO==0.7.1; sys_platform=="linux"
